Move axis orientation test to correct class
